requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.4
defusedxml==0.7.1
dj-database-url==1.0.0
Django==4.0.6
django-ranged-response==0.2.0
django-shopping-cart==0.1
django-simple-captcha==0.5.17
django-templated-mail==1.1.1
djangorestframework==3.13.1
docopt==0.6.2
idna==3.3
itypes==1.2.0
MarkupSafe==2.1.1
mysqlclient==2.1.1
oauthlib==3.2.0
Pillow==9.2.0
pipreqs==0.4.11
pycparser==2.21
PyJWT==2.4.0
python-decouple==3.6
python3-openid==3.2.0
pytz==2022.1
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.2
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.11
yarg==0.1.9
